GOCBC-1032: Add internal function for getting the status of a bucket capability
Motivation
----------
We recently added an internal function for getting the current
status of a given capability. This would be better if the function
just returned the given status for a capability. At present if the
user requests verification that a status is supported and the sdk
returns false then they do not know if the status is actually
unsupported or unknown so must check again with status unknown, at
which point the status could have changed to support so the verification
check would once again return false, misleadingly.

Changes
-------
Remove HasBucketCapabilityStatus function.
Add BucketCapabilityStatus function.

Change-Id: I9bd96826d3063dfb4d17871fd8538162abfe4c5a
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/146928
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>