GOCBC-1044: Add metrics interface and support
Motivation
----------
As a part of updating observatibility support we are adding support
for metrics. Gocbcore should provide a metrics interface and
also support the responses metric, that is a call to record latency
for value recording when we receive a response.

Changes
-------
Add metrics interface with default implementation of noop.
Add call to get a value recorder and record the latency when
we receive a response via HTTP or in memdclient.

Change-Id: I4cc5f7502484201545b00549547627076e6d90c9
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/149362
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>