GOCBC-1066: Check if poller controller is stopped before forcing http
Motivation
----------
We have an edge case which can lead to the poller controller calling
stop on one of the pollers twice, which leads to a panic. We should
not do that.

Changes
-------
Update ForceHTTPPoller to check if the poller controller is stopped
or the active controller is nil, and bail out if either are true.

Change-Id: Ia883e7dce038a15138e37a01248792ee44ebdd3c
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/147711
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>