GOCBC-1068: Change preempted config update log level to warn
Motivation
----------
When two configs are attempted to be applied at the same time we
reject one config and log an error. Whilst this is a case that
should not occur we should drop the log level to warn from error.
This happening can be indicative of something untoward happening
but in itself is not an error which causes the SDK to work incorrectly.

Changes
-------
Update preemptive config update log level to warn.

Change-Id: I3983c8c9a2cfcde1641d44569bf109e7f151327c
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/147909
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>