GOCBC-1071: Change get collection id packet to contain scope.col in the value
Motivation
----------
A bug in the server was recently discovered where the concatenation
of scope name with collection name could exceed the maximum key size
allowed by the server. For this reason the protocol is changing to
contain the concatenated names in the packet value.

Changes
-------
Update the packet we send for get collection id so that the
concatenation is in the value rather than the key.

Change-Id: I5769e72a135e9c5c0dc85d810046ca1955a9088b
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/148864
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>