GOCBC-1073: Cleaned up some buggy behaviours.
Motivation
----------
A recent change implemented to optimize some aspects of the client
caused an issues to appear with the ordering of channel sends during
shutdown as well as packets not flowing through the DCP callbacks
appropriately due to their Status code potentially being non-zero
when DCP messages were expected to have a zero status code.

Changes
-------
- Fixed a major issue where the packet pool would cause some DCP
events to be lost due to not resetting the packet to zero, and
instead they would trigger unexpected End() events.
- Modified the handling of the DCP processor shutdown to no longer
have a race that could lead to getting stuck.  This also fixes an
issue where a request to close a client would flush the internal
buffer before shutting down rather than shutting down immediately.
- Modified tryCallback handling to not attempt to call internalCancel
on a packet which we already know has been cancelled already.
- Fixed a minor inconsistency in opmap draining.

Change-Id: I001a61b9ad52102e8303df1ee011fd995f81bc87
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/149098
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>