GOCBC-1075: Return the correct error from buildAuthHandler
Motivation
-----------
Within buildAuthHandler we are returning the incorrect error
if callErr is not nil. This could lead to stange behaviour downstream
of the function call as both of the channels and the error object would
be nil. This could result in the callee waiting indefinitely on a
channel.

Changes
-------
Return the correct error from buildAuthHandler.

Change-Id: I2c7089ff98d69d280710b68da2d7e09ff8cca6e5
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/149326
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>