GOCBC-1079: Don't retry auth if error is request cancelled
Motivation
----------
On bootstrap if auth fails and the error isn't a valid
authentication error then we will retry with a different mechanism.
There's no point in doing this if bootstrap has been cancelled.

Changes
--------
Add check for if the error from auth request is request canceled
and if so then return the error.

Change-Id: I70a72de53c43d6f9cf360ffba5f70cbe94dd6d95
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/150458
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>