GOCBC-1080: Don't reconnect connections on first config seen supporting collections
Motivation
----------
When we see a config that has different collections support state
to the last config that we saw then we reconnect all of the connections.
This means that against 7.0 we are always reconnecting on first
config seen. We shouldn't do that, we should just perform a takeover.

Changes
-------
Add a check when applying configs in the kv muxer, if the old
mux state has a revid of -1 and collections are enabled then always
perform a takeover.

Change-Id: Ib42e49396801a1748c145da877aedbab500db9ba
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/150460
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>