GOCBC-1081: Don't immediately retry cccp if error is request cancelled
Motivation
----------
At the moment if a cccp request was cancelled then we retry
against the next node. Instead we should stop attempting cccp
for that iteration and allow the poller loop to loop back around,
where it'll probably exit.

Changes
-------
If a cccp request fails due to request cancelled then stop sending
cccp requests for that iteration and instead loop back around the
outer loop.

Change-Id: I2a19a9ea1c418c6ca1813011ddfbadc2795016ee
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/150463
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>