GOCBC-1082: Don't apply backoff for reconnectiong after request cancelled
Motivation
----------
When the server indicates that it now supports collections (and
they're enabled our side) when previously it didn't then we do a
SDK wide connections rebuild so that all connections have collections
enabled and supported. If one of those connections is in bootstrap
at this point then the dialler will see an error occurred during
bootstrap (we cancelled the operation) and will add the node to
the backoff cache so no further connections to that node will be
attempted for 5s.

Changes
-------
Do add the node to the server failures map if the error from
dialling the connection or bootstrap is a cancellation error.

Change-Id: Id12a789ead88d15d9b856cdc305f2df3c69c81fa
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/150993
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>