GOCBC-1062: Convert DCP interfaces to use structs
Motivation
---------
To support Point in Time Restore, support had to be added
for DCP snapshot marker v2.1. This meant the snapshotmarker
interface would need to be changed which would break backwards
compatibility.

Changes
------
To ensure compatibility is not broken in the future the DCP
interfaces now use structs which allows new attributes to be added
while honouring backwards compatibility.

The flags to enable Point in Time at a HELO level and a DCP stream
level have been added.

Change-Id: I3565cc0fdd12e0a0b1d803fa24cc3a9bf384241b
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/151158
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>