GOCBC-1077: Only return cas mismatch for n1ql if message matches
Motivation
----------
The query engine can return an error code 12009 for more cases
than just a cas mismatch. When we get a 12009 we need to check
the error message for "CAS mismatch". If it's a match then return
a cas mismatch but otherwise a DML failure.

Changes
-------
Add DMLFailure errors.
Add extra check for query 12009 error to check if the message
contains "CAS mismatch".

Change-Id: Ia95689ed1f301bd32369f5d08fad353f98d58b13
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/152909
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>