GOCBC-1093: Add support for revEpoch
Motivation
----------
The server team has added a new revEpoch field to bucket
configurations. The new field should be compared before revID
to see if the config should be applied. If the revEpoch of the
old config and new config match then we should compare the
rev ids as we did before. Otherwise, if the new revEpoch is lower
then ignore the config and if newer apply it.

Changes
-------
Add revEpoch to the config object.
Add revEpoch comparisons to the config manager.

Change-Id: I378c6405575ee02f0ed253a9e0a9ef34a8e933bc
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/152916
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>