GOCBC-1043: Don't append bucket name to capi endpoint in config
Motivation
----------
When building the config we take the capi endpoints and the bucket
name to them. This means that later we have to strip it off in the
diagnostics component. We should add the bucket name at the point
that we need the endpoint to have the bucket name (i.e. views requests).

Changes
-------
Update config parsing to not append the bucket to the capi endpoints.
Update http mux get capi endpoints to append the path escaped bucket
name to each endpoint.
Remove logic to strip bucket name from diagnostics component.

Change-Id: Ie92915119a75359d10b644dcb4b3d0b4892f4f0f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/152921
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>