GOCBC-1096: Add number of retries to CMD and HTTP spans
Motivation
----------
The observability RFC has now added a requirement for number of
retries to be placed on the outer level span. We cannot do that in
gocb, so in gocbcore we need to add retries to the CMD span and the
HTTP dispatch span.

Changes
--------
Add number of retries to CMD and HTTP dispatch spans.

Change-Id: Ic4d4a51403e43bb4df3c39c486cabbd8827c866f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/153458
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>