GOCBC-1102: Move context instantiation in http waituntilready
Motivation
----------
In WaitUntilReady if pingng one of the HTTP services fails then
WaitUntilReady if a ping fails then we call the context cancel
func, which lives for the lifetime of the WaitUntilReady call.

Changes
-------
Within checkHTTPReady wrap the HTTP request send and response
handling into a closure. Instantiate the context object within
the closure so that it only lives for the lifetime of that
specific waituntilready iteration.

Change-Id: I1a78692866f5338267c6d2e1953ce2ddc0154d07
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/154055
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/155481