GOCBC-1106: Allow hits field to be nil in search response
Motivation
----------
When we parse a fts response body we always expect hits to either
not to be in the body or to be an array. This is not always
the and hits can be present but set to nil.

Changes
-------
Update row streamer to optionally allow for "rows" to be set to
null.
Update fts to set the row streamer so that "hits" can be null.

Change-Id: Ibc273fcbbd7ea9b15a26c29d699faacbad02a8d6
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/155187
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/155554