GOCBC-1112: Parse top level search errors from error field
Motivation
----------
When fts replies with a non 200 status code then we parse the
error from the response body. We currently parse the error
from the status field which is always "fail", we should be
parsing from the error field.

Changes
-------
Update fts error response parsing to parse the error field instead
of the status field.

Change-Id: I83b8585ebc0af4bf1c5eb35ff9ef4026d7904abb
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/155201
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/155486