GOCBC-1127: Improve n1ql error handling
Motivation
----------
We have two places where we do not correct handle errors from
n1ql responses, we should correctly parse and expose these errors
to the user.

Changes
-------
Update old style prepared statements so that when we receive a 200
response for the prepare statement but there are no rows then we
attempt to parse the body and extract any errors.
Update the n1ql result Err function to create an error from the
error descriptions even if it is an error code that we don't know.

Change-Id: Iea84022fc5ac0bebabf825ac5af26ab082bf1049
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/155808
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>