GOCBC-1136: Cache metrics attributes maps
Motivation
----------
Benchmarking is showing that gocbcore HEAD is allocating a much higher
number of bytes per op than 9.1.5. This is primarily due to the fact
that we allocate a new map every operation to send attributes (service
and operation name) to the meter. We should cache these maps instead,
the map should remain quite small due to being limited by the number
of operations we actually expose.

Changes
-------
Add a metrics attributes sync map cache to tracer component and
use that when creating attributes.
Remove (move to tests) noop meter.
Don't use the cache or create any attributes when nil meter is
being used.

Change-Id: Id0873f1dff59d4e4569fa9e6be4fb57afb83e11c
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/157025
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>