Remove SetStatus from tracer interface
Motivation
----------
SetStatus isn't a committed part of the RTO RFC and marking it as
volatile would effectively make the tracer interface volatile. We
should remove it from the interface and if we want to add SetStatus
later we'll have to use a second interface and do type assertion.

Changes
-------
Remove SetStatus from the tracer interface.

Change-Id: I78d8a79fe8eae389eff7ab665763513480d68a98
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/157096
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>