GOCBC-1139: Use PLAIN auth mechanism with TLS by default for DCP
Motivation
----------
When using LDAP (and therefore TLS connections) with DCP and default
authentication mechanisms then connections will always fail. This is
because we will attempt to use SCRAM auth which the server does not
support for LDAP auth. The server will still advertise that it
supports SCRAM to us so we won't try other mechanisms which would
lead us to falling back to PLAIN auth.

We made a change to the standard agent so that when TLS is enabled
we only use PLAIN auth by default, we should apply this to the dcp
agent too.

Changes
--------
Only use PLAIN auth (by default) for DCP when TLS is set.

Change-Id: I6bf402e839dbf7ac0e8ed0373a561445202c8963
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/158720
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>