GOCBC-1147: Fix deadlock in kv error map
Motivation
-----------
On bootstrap e don't return error in the case that get error map fails.
This means that the error map channel is nil and we don't check for
that so we wait on the channel and hang forever.

Changes
-------
Add a check to ensure that the error map channel isn't nil before
trying to wait on it.

Change-Id: I7de780719d0f6eddc0561c30002e20eda776bce1
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/158868
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>