GOCBC-1148: Add ability to force reconnect all connections
Motivation
----------
Services like fts require a way to force all the SDK connections
to be rebuilt. This is required for cases such as when the user
changes the root certificate, the server does not kill the connections
when this happens so it needs to happen SDK side.

Changes
-------
Add a ForceReconnect function to agent and dcpagent.
Wire in the force reconnect to the kv muxer which will create a new
mux state based on the current mux state, close all old pipelines and
connect new ones.
Add the ability to pass an error down through the Close functions
into memdclient so that we can specify the error that should be
returned for any in flight requests.

Change-Id: I2ba26bc31cbeb35e88d84918c7a3ec99a0b91b96
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/158992
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>