GOCBC-1146: Add support for user impersonation to n1ql, analytics and fts
Motivation
----------
We already added support for user impersonation to kv and we also
need to add it for n1ql, fts, and analytics.

Changes
-------
Add a User field to each of n1ql, fts, and analytics options.
Add a User field to httpRequest and HTTPRequest.
If User is set on the request then add the cb-on-behalf header to
the http request.

Change-Id: Ib4fc45b9a00acd1734d5a8124d0031c3c486f09a
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/158906
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>