GOCBC-1139: Use PLAIN auth mechanism with TLS by default for DCP
Motivation
----------
When using LDAP (and therefore TLS connections) with DCP and default
authentication mechanisms then connections will always fail. This is
because we will attempt to use SCRAM auth which the server does not
support for LDAP auth.

We made a change to the standard agent so that when TLS is enabled
we only use PLAIN auth by default, we should apply this to the dcp
agent too.

Changes
--------
Only use PLAIN auth for DCP when TLS is set.

Change-Id: Iccc06a79e5c8dbcbdcf563113074fe10b9fbd6d8
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/158937
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>