GOCBC-1162: Add feature for initially connecting without TLS when using TLS
Motivation
----------
For localhost connections ns_server supplies server teams like FTS
the local address as 127.0.0.1 but supplies the certificate as the
external IP for the node. This means that when UseTLS is set to true
and the service tries to initially bootstrap then connections fail
because the certificate does not match the address being used.

We should add a way that allows UseTLS to be set but performs the
initial bootstrap without using TLS, that way we will establish the
initial connection, fetch a config, and then (re)build all the
connections using TLS.

Changes
--------
Add an option to SecurityOptions to allow initially connecting
over nonTLS when TLS is set to be used.
When the option is set then do not set the TLS config when dialling
memcached connections when a cluster config hasn't been seen yet.
When the option is set then seed the http addresses with HTTP
protocol rather than HTTPS.

Change-Id: I99f08c245d13e04830567853bfead39bc606f23f
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/160114
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>