GOCBC-1155: Add extra info to connection fail log messages
Motivation
----------
When the SDK cannot connect then we log a message such as
"connection refused" which is returned to us by the Dial function.
This does not give the user much information with where to try
look to fix the problem. We should enhance the error message with
a bit of information with where users could try looking.

Changes
-------
If dialMemdConn returns an error and it isn't a cancelation error
then enhance the error with the message "check server ports and
cluster encryption setting" which will then be logged.

Change-Id: Ie27f28ec077550d5933d4c772e55e4732b4e86e5
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/160894
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>