GOCBC-1169: Return nil if NextRow called on streamer after finalize
Motivation
----------
If a user repeatedly calls NextRow on a query streamer (e.g. for
n1ql) then when there are no more rows we finalize the streamer,
which includes setting the row streamer to nil. This means that
if the user calls NextRow again on the query streamer then it will
cause a panic. We should just return nil instead.

Changes
-------
If NextRow called on query streamer after finalize then return nil.

Change-Id: Iac9efcd8a1fe025282458209e244a0d473e0fc25
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/160971
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>