GOCBC-1177: Be more selective on when to apply node dial backoff
Motivation
----------
It's possible for the sdk to establish a
connection to the server but during bootstrap for the server to
terminate the connection. When this happens the SDK can get into a
state where the memdclient is already closed and tries to dispatch
one of the bootstrap operations.

This leads to the memdclient rejecting the operation with a request
cancelled error. If this error is returned for a bootstrap critical
operation (e.g. auth) then bootstrap is failed and the error
returned to the memdclient dialler.

If bootstrap fails against a node then the dialler will wait a set
amount of time before re-attempting bootstrap against that node.
This wait does not happen when bootstrap fails with a request
cancelled error, like returned in the scenario above. This leads
the SDK to immediately retry bootstrap, which can lead to all this
retriggering from the top and leading to a loop of bootstrapping.

An effect of this is that we create a large number of buffered
readers for connections in rapid succession, each 20MB. This can
lead to an explosion in the memory usage of the SDK.

Changes
-------
Change forcibly reconnecting connections to always use
errForcedReconnect.
Add new error types which are more descriptive of why memdclient
has rejected an operation.
Always apply node level bootstrap backoff unless the error is forced
reconnect.

Change-Id: I5e4a6fbe59c90f593c3e938d1521f96f129660b2
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/162906
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>