GOCBC-1178: Don't remove the poller controller watcher from config updates
Motivation
----------
Within the poller controller we have a watcher for when the SDK
receives a cluster config. The watcher is responsible for
transitioning from using the http poller to the cccp poller if
relevant. However we remove this watcher as soon as we see a
config.

There are scenarios where we can end up with the http poller being
triggered during normal operation of the SDK, not just at startup
time. This can occur when a node is being rebalanced out of the
cluster and the timing is just right that a node is bootstrapping
against it and that node replies to select bucket with a bucket not
found error.

Due to this we should leave the watcher running so that we can
always transition back to CCCP again.

Changes
--------
Only remove the poller controller watcher from config updates if
the bucket type is memcached.

Change-Id: I7f09577149c84e3495bd0ece0ed2e53d05e0daee
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/163069
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>