GOCBC-1179: Gracefully close memdclients on pipeline disconnect/reconnect
Motivation
----------

At present within the ForceConnect call (which reconnects all
connections for purposes of picking up new certificates) any in
flight operations are terminated with an ErrForcedReconnect. This
is not great, and this logic is actually shared by what happens
when the SDK becomes aware that the cluster has collections support
when previously it didn't - the difference being that in flight
requests are terminated with io.EOF.

What should actually happen is that we create new connections but
allow the old connections to stay live until either they get
terminated or their op list empties (i.e. all in flight requests
are completed).

Changes
-------
Add a GracefulClose function to memdclient.
After graceful close has been triggered check the size of the
op list when we receive responses from the server. If the list
is empty then close the client.
Update memdpipelineclient Close to not close the memdclient,
instead returning it to the caller.
Add a GracefulClose function to memdpipeline which closes the
pipeline clients, fetching but not closing the memdclients and
returning them upstream.
Update memdpipeline Close to close the memdclients returned by
pipeline client close.
Update kvmux so that on pipeline shutdown (outside of kvmux close)
it spins up a new goroutine per memdclient returned.
Inside this goroutine monitor for the memdclient shutting down
and also the kvmux shutting down.
If the kvmux shuts down before the memdclient then forcibly close
the memdclient.

Change-Id: I3518aacb284e98f195dc9e9cba7ec28f236abb00
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/163084
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>