GOCBC-1180: Add support for ns_server scheme and seed poller
Motivation
-----------
In order to support requirements of server components we should add
a new connection string scheme of ns_server. When used we will
allow only a single seed address in the connection string. This
single seed address will then be used for ALL config fetches using
a special config poller and ALL connections to this seed will be
made over non-SSL (as such this should really only ever be used
with localhost but for the purposes of testing we will allow other
addresses). All other connections will made over SSL unless the
UseTLS config option is overidden.

Changes
-------
Add a new ns_server connection string scheme.
When ns_server is used set TLS to true and set a new NoTLSSeedNode
agent config option to true.
Add a new poller that only polls against the seed node, and
satisfies the same interface as the poller controller - taking its
place when used.
Update internal endpoint representations to be a new routeEndpoint
type which contains both address and whether or not the endpoint
is the encrypted endpoint.
Update config parsing so that when parsing the block owned by the
node that was used to fetch the config then SSL ports are never
used.

Change-Id: I29aab44b27fd7a4547cb4586cc7e5cec6f588ee8
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/163207
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>