GOCBC-1181: Add ReconfigureSecurity function
Motivation
------------
Add a new ReconfigureSecurity function which will allow users to
change TLS options such as whether TLS is on or off and maybe
change the AuthProvider. This function will only be allowed to be
used when the ns_server scheme is in use.

The function will implicitly trigger a SDK wide connection reset
via a forced config fetch.

Changes
-------
Add ReconfigureSecurity functions to Agent and DCPAgent.
Update config manager to include both secure and insecure endpoints
in the route config that it pushes to other components.
In the relevant muxers use the current mux state to decide which
endpoints to use from the config at mux state creation time.
Push tls config, auth and auth mechanisms into the kvmux and
httpmux when ReconfigureSecurity is called.
Pass tls config, auth and auth mechasisms from the kvmux to the
dialler at connection time.
Update http client to use a callback into the http mux to decide
whether or not to use tls.

Change-Id: I8fa8febff3bdcd380a59f7c8ed79c8c2f3b0c4ea
Reviewed-on: http://review.couchbase.org/c/gocbcore/+/163482
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>