GOCBC-1190: Add functionality to send queries to specific endpoints
Motivation
----------
A requirement of query in transactions is that all queries in a
transaction must go to the same node. We need to add a way to
facilitate doing that.

Changes
-------
Add endpoint to n1ql row reader and n1ql options.
Wire endpoint into the existing field on http request/response.

Change-Id: I3fc6d473caa81fd6e2677c7e8c7732e2ac6a33aa
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/165075
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>