GOCBC-1193: Add ResponseBody field to N1qlError
Motivation
-----------
When we can't parse the response body as a jsonN1QLErrorResponse
then we will just return the N1QLError as "query error". We should
not just swallow any of the error body meaning that the user a)
cannot debug as easily and b) cannot then access the body if they
need it. This is true for the transactions body which needs to be
able to access the response body and parse it even though it is
not a jsonN1QLErrorResponse.

Changes
--------
Add a ResponseBody field to N1qlError which is populated with the
body from the n1ql response.

Change-Id: I0478864b22be131f284baa3fde1d165d1e0a19d3
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/165092
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>