GOCBC-1194: Check all bootstrap nodes for default network type first
Motivation
-----------
When we build the first route config we should check all of the
bootstrap nodes to see if they have an entry in the default
network for alternate addresses before trying to build an external
config.

Changes
--------
Move the external network type check to occur after all bootstrap
nodes have been checked for entries against the default network
type.

Change-Id: I73b0550217a2479f0ccb30b737adb0f2b8513783
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/165757
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>