GOCBC-1196: Add error text and status code to http service errors
Motivation
----------
In order to improve debuggability and ux we should add the error
text and status code to analytics, query, and view errors
where we can when we get an error that we cannot parse into a
typed error that we know about. Search errors already have these
fields.

Changes
-------
Add error text and status code to analytics, query, and
view errors.
Add plumbing to get the error text and status code into the
errors where applicable.

Change-Id: I41858243c3b50a1404ef1c79340c6913147b2072
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/166339
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>