GOCBC-1205: Complete telemetry on dispatch error and op timeout
Motivation
----------
When an operation fails to be dispatched or is timed out we are
not finishing the span created for the op. We are using
defer for http based operations so those spans are finishing
early. We are also using deferred everywhere for metrics, this
means that metrics are always going to be a very low value.

Changes
-------
Create a new op telemetry type which wraps the opTracer
and stores details for recording metrics.
When finish is called on the op telemetry type finish
the tracer and record metrics.
Call finish when op times out and is cancelled successfully.
Don't use defer for spans and metrics on http based operations.

Change-Id: If2f1f8cef792fe9a33ceb359cec20fbb311bc3ef
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/166852
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>