GOCBC-1208: Don't fallback to http polling with no bucket
Motivation
----------
It has been observed that there is a scenario in which we are
running the http poller when the agent does not have a bucket.
This will never succeed and we should guard against it.

Changes
-------
Use callbacks to the agent level so that any place that we will
fallback to http polling must first check the bucket name
on the agent is not empty.
If bucket name is empty then don't create http poller.

Change-Id: Ic0ee2f953a98f3a8ac75f66021abf681a91cd7b9
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/167039
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>