GOCBC-1209: Update DCP agent to tell kv mux about NoTLSSeedNode
Motivation
----------
At present the ns server scheme is not working with the dcp agent.
The server rejects connections because we are sending requests
to the tls endpoint but not setting up any tls config. This is
because the DCP agent is not telling the mux whether NoTLSSeedNode
has been set.

Changes
-------
Update DCP agent to tell kv mux about NoTLSSeedNode

Change-Id: Ic9d00348c1b586e44f85ef5b33fd83026e72a5e9
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/167062
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>