Update TestCollectionsComponentCollectionsSupportedCollectionUpdate
Motivation
----------
TestCollectionsComponentCollectionsSupportedCollectionUpdate
intermittently fails because dispatch is called too many times.
This is likely because of a timing issue, we should use channels
to lockstep the behaviour that we want.

Changes
-------
Add a channel to the test which causes the initial collection
refresh to not return until after we have dispatched all of the
initial operations.

Change-Id: I4326d5bbec1a82a569caf357faff3bff4741e141
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/167563
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>