GOCBC-1214: Always use seed node address when NoTLSSeedNode used
Motivation
----------
When NoTLSSeedNode is set then we will fetch the cluster config
from the seed node over non-tls. If the seed node is set to
localhost then the cluster map can contain a different address
e.g. the actual IP for the node. When this happens we take the
address from the cluster map and then attempt to use it with
non-TLS connections. If the server has TLS mode set to strict then
this causes the SDK to be unable to connect (as only the loopback
address is allowed to use non-TLS).

Changes
-------
Pass the NoTLSSeedNode setup down at route config setup time.
If NoTLSSeedNode is set and the node is the seed node then
ignore the cluster map and set the address to the address that we
used to fetch the config.

Change-Id: I531948108ef6de915bc43ea2f68b5ff02a54653c
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/168427
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>