GOCBC-1221: Add support for improved query error handling
Motivation
----------
Query have added support for new fields in the error payload:
reason and retry. We should add support for these new fields.
We should also add a new retry strategy for indicating that
an error is triggering a retry because the query engine
indicated to do so.

Changes
-------
Add Reason and Retry to QueryErrorDesc.
Add enhanced handling for error code 12009 to interrogate the
reason block to check for more detailed error codes.
Add a new retry strategy for when we retry due to the error
indicating to do so.

Change-Id: Ia906dfdecdbc182705dfeb89690fc7f81e1843f5
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/169039
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>