GOCBC-1222: Use a sync.pool for memd.conn readers
Motivation
-----------
At present if we have a large number of connections per node (e.g.
a number of agents each with a kv pool size of say 8) and something
goes wrong and we end up in a reconnect loop for a couple of
minutes then we can up blowing up our memory footprint as we create
a 20MB buffer per connection.

We should move to using a pool so that we reuse the buffers.

Changes
-------
Add a sync.Pool to memd package for *bufio.Reader.
Use the pool for memd.Conn readers.
Add a Release function to memd.Conn and release the reader when
called.
Add a Release function to memdConn.
Call Release from memdclient, with added synchronization to ensure
that the memd.Conn ReadPacket has completed before Release is called.

Rudimentary performance testing using cbgoclibench suggests
negligible impact to throughput and cpu usage.

Change-Id: If93b53255f100560018f811ff167aa9fe3581de3
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/168871
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>