GOCBC-1224: Check if memdclient already closed in Close
Motivation
----------
When a node that we are connected to goes down the connection to
the SDK is severed. This causes the memdClient to internally do a
close, closing and releasing its memdConn. The memdpipelineclient
then loops back around and tries to fetch a new memdClient. The
dialer can manage to establish the connection but bootstrap then
fails, which leads it to call Close on the memdclient. The first
read that the memdClient attempts errors and that triggers the
client to internally do a close. The bootstrap error is returned to
the dialer which then closes the memdclient. This second close
causes the memdclient's memd.Conn to panic as the reader is already
set to nil.

Changes
-------
Add a check to see if the memdclient is already closed to Close.

Change-Id: I1a57c5f95ba238a7614663724e0f9863b5ac5e2b
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/169133
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>